migrate switcher.py to plugins\switcher\widgets.py
